Initial "look at move order" functionality
At the moment the rotation is "bounded" to the Y axis.
Meaning that the model will try to keep facing "up"
no matter the rotation.

An example will make this clear, assume we have a plane and the axes:

    ^ Y
    |
    + - > X

and the plane model has its origin at the back:

                 \
                 ====-
                 .
             ...............
                 .
                 \\/
                 .\\
                 . \\
                 .   \
                 .
                 .     Target
                 .
             ...............
                 .
               \//
               //.
              // .
             /   .
                 .
         Target  .
                 .
             ...............
                 .
          Target .
                 .
              \  .
              \\ .
               \\.
                \\--
                 .
             ...............
                 .
                 .     Target
                 .
                 .  /
                 . //
                 .//
               --//
                 .

Ignoring the lack of scale due to monospace fonts, in our example the
rudder will never face -Y.